chore: Update file paths in adr-tool.csproj
Changed the paths for `None` elements in `adr-tool.csproj`.
The path for `icons8-decision-96.png` was updated from `..\Assets\icons8-decision-96.png` to `..\..\Assets\icons8-decision-96.png`, and the path for `README.md` was changed from `..\README.md` to `..\..\README.md`.
This indicates that the files are now located in a higher directory than previously assumed.